♻️ :: UseCase 하나로 합침
